requirements.txt
Flask
RPi.GPIO